create common publication type to avoid circular dependencies when importing
