+++ src/pages/publication/contributors_tab/components/AffiliationsCell.tsx
@@ -4,7 +4,7 @@ import { Button } from '@material-ui/core';
 import SelectInstitution from '../../../../components/institution/SelectInstitution';
 import Modal from '../../../../components/Modal';
 import { useFormikContext, FormikProps } from 'formik';
-import { Publication, BackendTypeNames } from '../../../../types/publication.types';
+import { Publication } from '../../../../types/publication.types';
 import { SpecificContributorFieldNames } from '../../../../types/publicationFieldNames';
 import { FormikInstitutionUnit } from '../../../../types/institution.types';
 import { useTranslation } from 'react-i18next';
@@ -20,6 +20,7 @@ import { NotificationVariant } from '../../../../types/notification.types';
 import AffiliationHierarchy from '../../../../components/institution/AffiliationHierarchy';
 import Card from '../../../../components/Card';
 import NormalText from '../../../../components/NormalText';
+import { BackendTypeNames } from '../../../../types/publication_types/common.publication.types';
 
 const StyledCard = styled(Card)`
   display: flex;
