+++ src/types/contributor.types.ts
@@ -1,4 +1,5 @@
-import { BackendType, BackendTypeNames } from './publication.types';
+import { BackendType } from './publication.types';
+import { BackendTypeNames } from './publication_types/common.publication.types';
 
 // For available roles, see https://github.com/BIBSYSDEV/nva-datamodel-java/blob/develop/src/main/java/no/unit/nva/model/Role.java
 export enum ContributorRole {
