+++ src/types/customerInstitution.types.ts
@@ -1,4 +1,5 @@
-import { BackendType, BackendTypeNames } from './publication.types';
+import { BackendType } from './publication.types';
+import { BackendTypeNames } from './publication_types/common.publication.types';
 
 export interface CustomerInstitution extends BackendType {
   identifier: string;
