+++ src/types/project.types.ts
@@ -1,4 +1,5 @@
-import { BackendType, BackendTypeNames } from './publication.types';
+import { BackendType } from './publication.types';
+import { BackendTypeNames } from './publication_types/common.publication.types';
 
 export interface Project extends BackendType {
   id: string;
