+++ src/types/publication_types/journal.publication.types.ts
@@ -1,7 +1,8 @@
-import { BackendType, PagesRange, BackendTypeNames, emptyDate, PublicationDate } from '../publication.types';
+import { BackendType, PagesRange, PublicationDate } from '../publication.types';
 import { PublicationType, JournalType } from '../publicationFieldNames';
 import { LanguageValues } from '../language.types';
 import { Contributor } from '../contributor.types';
+import { BackendTypeNames, emptyDate } from './common.publication.types';
 
 interface JournalPublicationInstance {
   type: JournalType | '';
