+++ src/types/publication_types/report.publication.types.ts
@@ -1,7 +1,8 @@
-import { BackendType, BackendTypeNames, emptyDate, PublicationDate, PagesMonograph } from '../publication.types';
-import { PublicationType, JournalType, DegreeType } from '../publicationFieldNames';
+import { BackendType, PublicationDate, PagesMonograph } from '../publication.types';
+import { PublicationType, DegreeType } from '../publicationFieldNames';
 import { LanguageValues } from '../language.types';
 import { Contributor } from '../contributor.types';
+import { BackendTypeNames, emptyDate } from './common.publication.types';
 
 interface ReportPublicationInstance {
   type: DegreeType | '';
